Refactor imports and remove unused code
